added ability to add pack remote entries to RO enabled packs
git-svn-id: svn+ssh://rubyforge.org/var/svn/myexperiment/branches/packs@3616 4f7b4649-9ca9-420a-81b3-55a188e4d46f

+++ app/models/resource.rb
@@ -149,8 +149,10 @@ class Resource < ActiveRecord::Base
 
     new_description = create_rdf_xml { |graph| graph << description }
 
-    content_blob.destroy if content_blob
-    update_attribute(:content_blob, ContentBlob.new(:data => new_description))
+    unless is_resource
+      content_blob.destroy if content_blob
+      update_attribute(:content_blob, ContentBlob.new(:data => new_description))
+    end
   end
 
   def annotations
